requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
# streamlit
# langchain
# langchain-community
# langchain-core
# ctransformers
# sentence-transformers
# torch
# uvicorn
# python-box